Add files via upload
1. Data Wrangling: This is the first step where inspection of data is done to make sure NULL values and missing values are detected and data replacement methods are used to replace missing or NULL values.
2. Add a new column named Time of day to give insight of sales in the Morning, Afternoon and Evening. 
Add a new column named Day name that contains the extracted days of the week on which the given transaction took place (Mon, Tue, Wed, Thur, Fri). 
Add a new column named Month name that contains the extracted months of the year on which the given transaction took place (Jan, Feb, Mar).